Feature/git dirty repos

Add a git "module" that checks if workspace repositories are clean or not.

If not, asks if user wishes to continue with consequence of potentially loosing local files/commits/modifications.

See merge request !5

+++ drush.py
@@ -20,11 +20,14 @@ from __future__ import unicode_literals
 from fabric.api import task, roles, env
 from fabric.contrib.console import confirm
 from fabric.colors import red, green
+from fabric.utils import abort
 
 from datetime import datetime
 
 import helpers as h
 import core as c
+
+from git import isGitDirty
 
 
 @task(alias='make')
@@ -34,6 +37,12 @@ def make(action='install'):
     Build the platform by running the Makefile specified in the local_vars.py configuration file.
     """
 
+    if env.get('interactive_mode', True):
+      if (isGitDirty()):
+        if (not confirm(red('There are warnings on status of your repositories. '
+                        'Do you want to continue and reset all changes to remote repositories'' states?'), default=False)):
+          abort('Aborting "drush {}" since there might be a risk of loosing local data.'.format(action))
+
     drush_opts = "--prepare-install " if action != 'update' else ''
 
     # Update profile codebase
@@ -41,19 +50,18 @@ def make(action='install'):
         drush_opts += "--contrib-destination=profiles/{} ".format(env.site_profile)
         h.update_profile()
 
-    if env.get('interactive_mode', True) == 'false':
+    if not env.get('interactive_mode', True):
         drush_opts += "--translations=fr "
     elif confirm(red('Say [Y] to {} the site at {} with the French translation, if you say [n] '
                      'the site will be installed in English only'.format(action, env.site_root))):
         drush_opts += "--translations=fr "
 
-    if env.get('interactive_mode', True) != 'false':
+    if env.get('interactive_mode', True):
         drush_opts += " --working-copy --no-gitinfofile"
     if not h.fab_exists('local', env.site_root):
         h.fab_run('local', "mkdir {}".format(env.site_root))
     with h.fab_cd('local', env.site_root):
         h.fab_run('local', 'drush make {} {} -y'.format(drush_opts, env.makefile))
-
 
 @task
 @roles('local')
